requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
pymongo==4.10.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.19
jinja2==3.1.4
email-validator==2.2.0
